Implement bytecode prevention across application modules and restructure UI components into a modular format; add new UI definitions and cleanup scripts.
