+++ server.py
@@ -1,2 +1,6 @@
 
+import sys
+# Prevent writing .pyc files during runs
+sys.dont_write_bytecode = True
+
 from server_scripts.server_main import server_function
