+++ server_scripts/helpers/__init__.py
@@ -1,3 +1,12 @@
+"""helpers package initializer.
+
+Disable writing bytecode files during import of helper modules.
+"""
+import sys
+import os
+
+sys.dont_write_bytecode = True
+os.environ.setdefault("PYTHONDONTWRITEBYTECODE", "1")
 # This file marks the helpers directory as a Python package after moving.
 import sys
 import os
