+++ ui_scripts/__init__.py
@@ -1,1 +1,9 @@
-# UI package for modularized components
+"""UI package initializer.
+
+Disable writing bytecode files when importing UI modules.
+"""
+import sys
+import os
+
+sys.dont_write_bytecode = True
+os.environ.setdefault("PYTHONDONTWRITEBYTECODE", "1")
