+++ ui_scripts/components/__init__.py
@@ -1,20 +1,1 @@
-"""components package for reusable UI helpers"""
-from .common_ui import (
-    nav_panel,
-    card_with_header,
-    card_body,
-    action_button,
-    download_button,
-    file_input,
-    get_custom_head,
-)
-
-__all__ = [
-    "nav_panel",
-    "card_with_header",
-    "card_body",
-    "action_button",
-    "download_button",
-    "file_input",
-    "get_custom_head",
-]
+# UI package for modularized components
