+++ ui_scripts/components/forecast_tab.py
