+++ ui_scripts/components/summary_tab.py
